requirements.txt
speedtest-cli
tk
